clips-mip-scheduler: CONTINUOUS instead INTEGER time vars
A very strange problem happens here. The comparison when constructing the
schedule-resource such that the events are arranged by their start time
from smaller to bigger...
The float comparison fails sometimes with no known reason comparing an
event starting at 0 with an event starting at a later time.
This is such a strange issue which all the other sides had been rolled
out. The problem is clearly that the float comparison does not always
give the expected result. I suppose this might has something to do with
float point representation in different  langs

+++ src/plugins/clips-mip-scheduler/clips_mip_scheduler_thread.h
@@ -74,7 +74,7 @@ private:
 			name = n;
 		};
 		std::string                name;
-		int                        duration = 0;
+		float                      duration = 0;
 		float                      lbound   = 0;
 		float                      ubound   = GRB_INFINITY;
 		std::map<std::string, int> resources;
@@ -97,7 +97,7 @@ private:
 	std::map<std::string, std::map<std::string, std::map<std::string, GRBVar>>> gurobi_vars_sequence_;
 
 private:
-	void set_event_duration(std::string env_name, std::string event_name, int duration);
+	void set_event_duration(std::string env_name, std::string event_name, float duration);
 	void set_event_bounds(std::string env_name, std::string event_name, float lb, float ub);
 	void
 	     add_event_resource(std::string env_name, std::string event_name, std::string res_name, int req);
